requirements.txt
django>=4.2
djangorestframework>=3.14
pandas>=2.0
django-cors-headers>=4.0 